1.0

+++ src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BidObjectMapper.java
@@ -39,7 +39,7 @@ public interface BidObjectMapper {
     /**
      * 根据biddeeId查询记录
      */
-    double countAmountByBid(Integer user_id);
+    Long countAmountByBid(Integer user_id);
     /**
      * 根据user_id查询招标项目招标中的数量
      */
